Added some minor stuff
News are now described. Added search for product names in products and
message that product is added to cart. Products are now with rarity and
type not rarity and type id's. Comments are now with username not user
id. Added quantity to products in cart. Error if user is not found in
database when logging in. When registering an existing user, shows
validation error, that user already exists in database.